[refactor #108] 코드 리팩토링

+++ services/queue-manage/src/main/java/com/ticketPing/queue_manage/domain/repository/WorkingQueueRepositoryImpl.java
@@ -1,12 +1,11 @@
 package com.ticketPing.queue_manage.domain.repository;
 
-import static com.ticketPing.queue_manage.domain.model.WorkingQueueToken.tokenWithValidUntil;
 import static com.ticketPing.queue_manage.infrastructure.utils.TTLConverter.toLocalDateTime;
 
-import com.ticketPing.queue_manage.domain.command.workingQueue.DequeueWorkingTokenCommand;
-import com.ticketPing.queue_manage.domain.command.workingQueue.EnqueueWorkingTokenCommand;
+import com.ticketPing.queue_manage.domain.command.workingQueue.DeleteWorkingTokenCommand;
+import com.ticketPing.queue_manage.domain.command.workingQueue.InsertWorkingTokenCommand;
 import com.ticketPing.queue_manage.domain.command.workingQueue.CountAvailableSlotsCommand;
-import com.ticketPing.queue_manage.domain.command.workingQueue.RetrieveWorkingTokenCommand;
+import com.ticketPing.queue_manage.domain.command.workingQueue.FindWorkingTokenCommand;
 import com.ticketPing.queue_manage.domain.model.AvailableSlots;
 import com.ticketPing.queue_manage.domain.model.WorkingQueueToken;
 import java.util.Optional;
@@ -30,7 +29,7 @@ public class WorkingQueueRepositoryImpl implements WorkingQueueRepository {
     }
 
     @Override
-    public void enqueueWorkingToken(EnqueueWorkingTokenCommand command) {
+    public void insertWorkingToken(InsertWorkingTokenCommand command) {
         if (!isExistToken(command.getTokenValue())) {
             cacheToken(command);
             increaseCounter(command);
@@ -41,38 +40,55 @@ public class WorkingQueueRepositoryImpl implements WorkingQueueRepository {
         return redissonClient.getBucket(tokenValue).get() != null;
     }
 
-    private void cacheToken(EnqueueWorkingTokenCommand command) {
+    private void cacheToken(InsertWorkingTokenCommand command) {
         RBucket<String> bucket = redissonClient.getBucket(command.getTokenValue());
         bucket.set(command.getValue(), command.getTtl(), TimeUnit.MINUTES);
     }
 
-    private void increaseCounter(EnqueueWorkingTokenCommand command) {
+    private void increaseCounter(InsertWorkingTokenCommand command) {
         redissonClient.getAtomicLong(command.getPerformanceId()).incrementAndGet();
     }
 
     @Override
-    public Optional<WorkingQueueToken> retrieveWorkingToken(RetrieveWorkingTokenCommand command) {
+    public Optional<WorkingQueueToken> findWorkingToken(FindWorkingTokenCommand command) {
         RBucket<String> bucket = redissonClient.getBucket(command.getTokenValue());
         String value = bucket.get();
         if (value == null) {
             return Optional.empty();
         }
         long ttl = bucket.remainTimeToLive();
-        WorkingQueueToken token = tokenWithValidUntil(command.getUserId(), command.getPerformanceId(), command.getTokenValue(), toLocalDateTime(ttl));
+        WorkingQueueToken token = WorkingQueueToken.withValidUntil(command.getUserId(), command.getPerformanceId(), command.getTokenValue(), toLocalDateTime(ttl));
         return Optional.of(token);
     }
 
     @Override
-    public void dequeueWorkingToken(DequeueWorkingTokenCommand command) {
-        deleteToken(command);
+    public void deleteWorkingToken(DeleteWorkingTokenCommand command) {
+        switch (command.getDeleteCase()) {
+            case TokenExpired:
+                handleTokenExpired(command);
+                break;
+            case OrderCompleted:
+                handleOrderCompleted(command);
+                break;
+        }
+    }
+
+    private void handleTokenExpired(DeleteWorkingTokenCommand command) {
         decreaseCounter(command);
     }
 
-    private void deleteToken(DequeueWorkingTokenCommand command) {
+    private void handleOrderCompleted(DeleteWorkingTokenCommand command) {
+        if (isExistToken(command.getTokenValue())) {
+            deleteToken(command);
+            decreaseCounter(command);
+        }
+    }
+
+    private void deleteToken(DeleteWorkingTokenCommand command) {
         redissonClient.getBucket(command.getTokenValue()).delete();
     }
 
-    private void decreaseCounter(DequeueWorkingTokenCommand command) {
+    private void decreaseCounter(DeleteWorkingTokenCommand command) {
         redissonClient.getAtomicLong(command.getPerformanceId()).decrementAndGet();
     }
